Documentation update to include thumbnail aliases functionality.
The README.rst has been extended to be much more comprehensive
eventually with the idea of becoming the master documentation. Added
NOTE statements to files in the `docs` folder to refer to the
README.rst for up to date information.

Consolidating the documentation into one README.rst file simplifies
future maintainability and ought to be suffice for thsi project; as a
side bonus Github display README.rst by default assists with SEO.