Catch a potential OSError when getting EXIF info
_getexif is pretty touchy about errors happening. It was already wrapped
in a generous try/except which catches most issues. It looks like this
one is happening with the TIFF plugin (which is odd, since the image
does seem to be a JPEG - I guess PIL is just confused).

Oh well, just add another exception clause to the except;pass branch.

Fixes #151.